requirements.txt
python-telegram-bot==22.3
google-generativeai
python-dotenv
